fix visited url comparison in sub-task-agent
